pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Joshua Wallert", email = "jviper32@users.noreply.github.com"},
]
dependencies = [
    "lxml>=4.9.3",
    "requests>=2.31.0",
    "pqdm>=0.2.0",
    "deepgram-sdk>=2.10.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

